Moving the PipelineLoader from base to pipeline
